# swagger.yaml
swagger: "2.0"
info:
  title: Items API
  description: A simple inventory management API with GitHub OAuth login and API key protection
  version: "1.0"

basePath: /
consumes:
  - application/json
produces:
  - application/json

# ---------------------------------------------------------
# API KEY SECURITY
# ---------------------------------------------------------
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: X-API-Key
    description: >
      All protected routes require an API key.
      Include it in the request header as:
      X-API-Key: your-api-key

# Apply API key globally
security:
  - ApiKeyAuth: []

# ---------------------------------------------------------
# DATA MODELS
# ---------------------------------------------------------
definitions:

  Item:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: sword
      quantity:
        type: integer
        example: 3

  ItemInput:
    type: object
    required:
      - name
      - quantity
    properties:
      name:
        type: string
        example: sword
      quantity:
        type: integer
        minimum: 0
        example: 3

  ItemsResponse:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: "#/definitions/Item"
      total:
        type: integer
        example: 2

  OAuthResponse:
    type: object
    properties:
      message:
        type: string
        example: OAuth2 Authentication Successful
      github_user:
        type: string
        example: QuantumDevPro

  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: Unauthorized. invalid or missing API key

  SuccessResponse:
    type: object
    properties:
      message:
        type: string
        example: Item deleted successfully


# ---------------------------------------------------------
# API ROUTES
# ---------------------------------------------------------
paths:

  /:
    get:
      summary: API health check
      description: Returns API status.
      operationId: healthCheck
      security: []
      responses:
        200:
          description: API is running
          schema:
            type: object
            properties:
              message:
                type: string
                example: Secure API is running!
              status:
                type: string
                example: ok


  /login:
    get:
      summary: GitHub OAuth login
      description: Redirects the user to GitHub for authentication.
      operationId: githubLogin
      security: []
      responses:
        302:
          description: Redirect to GitHub OAuth login page

  /items:
    get:
      summary: Get all items
      description: Returns all items in the inventory.
      operationId: getItems
      responses:
        200:
          description: List of items
          schema:
            $ref: "#/definitions/ItemsResponse"
        401:
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"

    post:
      summary: Create item
      description: Adds a new item to the inventory.
      operationId: createItem
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/ItemInput"
      responses:
        201:
          description: Item created
          schema:
            $ref: "#/definitions/Item"
        400:
          description: Validation error
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"


  /items/{item_id}:
    get:
      summary: Get a single item
      description: Returns a specific item by ID.
      operationId: getItem
      parameters:
        - in: path
          name: item_id
          required: true
          type: integer
          description: Item ID
      responses:
        200:
          description: Item found
          schema:
            $ref: "#/definitions/Item"
        404:
          description: Item not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      summary: Update item
      description: Updates an existing item.
      operationId: updateItem
      parameters:
        - in: path
          name: item_id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/ItemInput"
      responses:
        200:
          description: Item updated
          schema:
            $ref: "#/definitions/Item"
        404:
          description: Item not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      summary: Delete item
      description: Removes an item from the inventory.
      operationId: deleteItem
      parameters:
        - in: path
          name: item_id
          required: true
          type: integer
      responses:
        200:
          description: Item deleted
          schema:
            $ref: "#/definitions/SuccessResponse"
        404:
          description: Item not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"